tasks/sancta/requirements.txt
pillow
qrcode[pil]
